initialize fake objects for go tests
