+++ pkg/utils/init.go
@@ -19,69 +19,63 @@ package utils
 import (
 	"fmt"
 	"os"
-	"path/filepath"
+	"sync/atomic"
 
 	direct "github.com/minio/direct-csi/pkg/clientset"
 	directcsi "github.com/minio/direct-csi/pkg/clientset/typed/direct.csi.min.io/v1beta2"
+	directcsifake "github.com/minio/direct-csi/pkg/clientset/typed/direct.csi.min.io/v1beta2/fake"
 
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/metadata"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/restmapper"
-	"k8s.io/client-go/tools/clientcmd"
-
-	"github.com/spf13/viper"
 
 	"k8s.io/klog/v2"
 )
 
-var directCSIClient directcsi.DirectV1beta2Interface
-var directClientset direct.Interface
-var kubeClient kubernetes.Interface
-var crdClient apiextensions.CustomResourceDefinitionInterface
-var apiextensionsClient apiextensions.ApiextensionsV1Interface
-var discoveryClient discovery.DiscoveryInterface
-var metadataClient metadata.Interface
+const MaxThreadCount = 40
 
 var (
-	initialized = false
+	initialized         int32
+	kubeClient          kubernetes.Interface
+	directCSIClient     directcsi.DirectV1beta2Interface
+	directClientset     direct.Interface
+	apiextensionsClient apiextensions.ApiextensionsV1Interface
+	crdClient           apiextensions.CustomResourceDefinitionInterface
+	discoveryClient     discovery.DiscoveryInterface
+	metadataClient      metadata.Interface
 )
 
-const (
-	MaxThreadCount = 40
-)
-
-func getKubeConfig() (*rest.Config, string, error) {
-	kubeConfig := viper.GetString("kubeconfig")
-	if kubeConfig == "" {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			klog.Infof("could not find home dir: %v", err)
-		}
-		kubeConfig = filepath.Join(home, ".kube", "config")
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", kubeConfig)
-	if err != nil {
-		if config, err = rest.InClusterConfig(); err != nil {
-			return nil, "", err
-		}
-	}
-	config.QPS = float32(MaxThreadCount / 2)
-	config.Burst = MaxThreadCount
-	return config, "", nil
+func GetKubeClient() kubernetes.Interface {
+	return kubeClient
 }
 
-func GetKubeConfig() (*rest.Config, error) {
-	config, _, err := getKubeConfig()
-	return config, err
+func GetDirectCSIClient() directcsi.DirectV1beta2Interface {
+	return directCSIClient
+}
+
+func GetDirectClientset() direct.Interface {
+	return directClientset
+}
+
+func GetAPIExtensionsClient() apiextensions.ApiextensionsV1Interface {
+	return apiextensionsClient
+}
+
+func GetCRDClient() apiextensions.CustomResourceDefinitionInterface {
+	return crdClient
+}
+
+func GetDiscoveryClient() discovery.DiscoveryInterface {
+	return discoveryClient
+}
+
+func GetMetadataClient() metadata.Interface {
+	return metadataClient
 }
 
 func Init() {
-	if initialized {
+	if atomic.AddInt32(&initialized, 1) != 0 {
 		return
 	}
 
@@ -128,32 +122,8 @@ func Init() {
 		fmt.Printf("%s: could not initialize metadata client: err=%v\n", Bold("Error"), err)
 		os.Exit(1)
 	}
-
-	initialized = true
 }
 
-func GetGroupKindVersions(group, kind string, versions ...string) (*schema.GroupVersionKind, error) {
-	discoveryClient := GetDiscoveryClient()
-	apiGroupResources, err := restmapper.GetAPIGroupResources(discoveryClient)
-	if err != nil {
-		klog.Errorf("could not obtain API group resources: %v", err)
-		return nil, err
-	}
-	restMapper := restmapper.NewDiscoveryRESTMapper(apiGroupResources)
-	gk := schema.GroupKind{
-		Group: group,
-		Kind:  kind,
-	}
-	mapper, err := restMapper.RESTMapping(gk, versions...)
-	if err != nil {
-		klog.Errorf("could not find valid restmapping: %v", err)
-		return nil, err
-	}
-
-	gvk := &schema.GroupVersionKind{
-		Group:   mapper.Resource.Group,
-		Version: mapper.Resource.Version,
-		Kind:    mapper.Resource.Resource,
-	}
-	return gvk, nil
+func SetDirectCSIClient(fakeClient *directcsifake.FakeDirectV1beta2) {
+	directCSIClient = fakeClient
 }
