+++ pkg/utils/init_test.go
@@ -1,0 +1,42 @@
+// This file is part of MinIO Direct CSI
+// Copyright (c) 2021 MinIO, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package utils
+
+import (
+	clientsetfake "github.com/minio/direct-csi/pkg/clientset/fake"
+	directcsifake "github.com/minio/direct-csi/pkg/clientset/typed/direct.csi.min.io/v1beta2/fake"
+
+	apiextensionsv1fake "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1/fake"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	runtime "k8s.io/apimachinery/pkg/runtime"
+	discoveryfake "k8s.io/client-go/discovery/fake"
+	kubernetesfake "k8s.io/client-go/kubernetes/fake"
+	metadatafake "k8s.io/client-go/metadata/fake"
+)
+
+func init() {
+	kubeClient = kubernetesfake.NewSimpleClientset()
+	directCSIClient = &directcsifake.FakeDirectV1beta2{}
+	directClientset = clientsetfake.NewSimpleClientset()
+	apiextensionsClient = &apiextensionsv1fake.FakeApiextensionsV1{}
+	crdClient = apiextensionsClient.CustomResourceDefinitions()
+	discoveryClient = &discoveryfake.FakeDiscovery{}
+
+	scheme := runtime.NewScheme()
+	metav1.AddMetaToScheme(scheme)
+	metadataClient = metadatafake.NewSimpleMetadataClient(scheme)
+}
