+++ pkg/utils/kubeutils.go
@@ -17,7 +17,16 @@
 package utils
 
 import (
+	"os"
+	"path/filepath"
+
+	"github.com/spf13/viper"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/restmapper"
+	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/klog/v2"
 )
 
 const (
@@ -314,4 +323,56 @@ func SanitizeFinalizers(finalizers []string) []string {
 		toRet = append(toRet, SanitizeLabelK(k))
 	}
 	return toRet
+}
+
+func getKubeConfig() (*rest.Config, string, error) {
+	kubeConfig := viper.GetString("kubeconfig")
+	if kubeConfig == "" {
+		home, err := os.UserHomeDir()
+		if err != nil {
+			klog.Infof("could not find home dir: %v", err)
+		}
+		kubeConfig = filepath.Join(home, ".kube", "config")
+	}
+
+	config, err := clientcmd.BuildConfigFromFlags("", kubeConfig)
+	if err != nil {
+		if config, err = rest.InClusterConfig(); err != nil {
+			return nil, "", err
+		}
+	}
+	config.QPS = float32(MaxThreadCount / 2)
+	config.Burst = MaxThreadCount
+	return config, "", nil
+}
+
+func GetKubeConfig() (*rest.Config, error) {
+	config, _, err := getKubeConfig()
+	return config, err
+}
+
+func GetGroupKindVersions(group, kind string, versions ...string) (*schema.GroupVersionKind, error) {
+	discoveryClient := GetDiscoveryClient()
+	apiGroupResources, err := restmapper.GetAPIGroupResources(discoveryClient)
+	if err != nil {
+		klog.Errorf("could not obtain API group resources: %v", err)
+		return nil, err
+	}
+	restMapper := restmapper.NewDiscoveryRESTMapper(apiGroupResources)
+	gk := schema.GroupKind{
+		Group: group,
+		Kind:  kind,
+	}
+	mapper, err := restMapper.RESTMapping(gk, versions...)
+	if err != nil {
+		klog.Errorf("could not find valid restmapping: %v", err)
+		return nil, err
+	}
+
+	gvk := &schema.GroupVersionKind{
+		Group:   mapper.Resource.Group,
+		Version: mapper.Resource.Version,
+		Kind:    mapper.Resource.Resource,
+	}
+	return gvk, nil
 }
